+++ message-format/src/main/java/de/sayayi/lib/message/util/MessageUtil.java
@@ -32,6 +32,133 @@ public final class MessageUtil
   private MessageUtil() {}
 
 
+  @Contract(value = "null, _ -> fail; _, _ -> param1", pure = true)
+  public static @NotNull String validateName(String name, @NotNull String nameDescription)
+  {
+    if (name == null)
+      throw new NullPointerException(nameDescription + " must not be null");
+    else if (name.isBlank())
+      throw new IllegalArgumentException(nameDescription + " must not be empty");
+
+    return name;
+  }
+
+
+  /**
+   * Determines if the specified character is a Unicode space character. A character is considered
+   * to be a space character if and only if it is specified to be a space character by the
+   * Unicode Standard. This method returns true if the character's general category type is any of
+   * the following:
+   * <ul>
+   * <li> {@code SPACE_SEPARATOR}
+   * <li> {@code PARAGRAPH_SEPARATOR}
+   * </ul>
+   *
+   * @param ch  the character to be tested
+   *
+   * @return  {@code true} if the character is a space character,
+   *          {@code false} otherwise
+   *
+   * @see Character#isSpaceChar(char)
+   */
+  @Contract(pure = true)
+  public static boolean isSpaceChar(char ch) {
+    return ((((1 << SPACE_SEPARATOR) | (1 << PARAGRAPH_SEPARATOR)) >> getType((int)ch)) & 1) != 0;
+  }
+
+
+  /**
+   * Returns a string without leading and trailing spaces. This method differs from
+   * {@link String#trim()} in that it doesn't trim newlines.
+   *
+   * @param s  string to trim, or {@code null}
+   *
+   * @return  trimmed string, or {@code null} if {@code s} is {@code null}
+   *
+   * @see #isSpaceChar(char)
+   */
+  @Contract(value = "null -> null; !null -> !null", pure = true)
+  @SuppressWarnings("DuplicatedCode")
+  public static String trimSpaces(String s)
+  {
+    if (s == null)
+      return null;
+
+    final var val = s.toCharArray();
+    var endIndex = val.length;
+    var startIdx = 0;
+
+    while(startIdx < endIndex && isSpaceChar(val[startIdx]))
+      startIdx++;
+
+    while(startIdx < endIndex && isSpaceChar(val[endIndex - 1]))
+      endIndex--;
+
+    return startIdx == endIndex
+        ? ""
+        : startIdx > 0 || endIndex < val.length
+            ? new String(val, startIdx, endIndex - startIdx)
+            : s;
+  }
+
+
+  /**
+   * Returns the trimmed variant of {@code s}. If {@code s = null} an empty string is returned.
+   *
+   * @param s  string to trim
+   *
+   * @return  trimmed string, never {@code null}
+   *
+   * @see #trimSpaces(String)
+   */
+  @Contract(pure = true)
+  public static @NotNull String trimSpacesNotNull(String s) {
+    return s == null ? "" : trimSpaces(s);
+  }
+
+
+  /**
+   * Tells if a string is empty.
+   * <p>
+   * A string is considered empty if it has zero length, or it contains spaces only.
+   *
+   * @param s  string, not {@code null}
+   *
+   * @return  {@code true} if the string is empty or contains spaces only, {@code false} otherwise
+   *
+   * @see #isSpaceChar(char)
+   */
+  @Contract(pure = true)
+  @SuppressWarnings("DuplicatedCode")
+  public static boolean isTrimmedEmpty(@NotNull String s)
+  {
+    final var val = s.toCharArray();
+    var endIndex = val.length;
+    var startIdx = 0;
+
+    while(startIdx < endIndex && isSpaceChar(val[startIdx]))
+      startIdx++;
+
+    while(startIdx < endIndex && isSpaceChar(val[endIndex - 1]))
+      endIndex--;
+
+    return startIdx == endIndex;
+  }
+
+
+  /**
+   * Returns whether string {@code s} is empty or not.
+   *
+   * @param s  string or {@code null}
+   *
+   * @return  {@code true} if the string is empty or {@code null}, {@code false} otherwise
+   */
+  @Contract(value = "null -> true", pure = true)
+  public static boolean isEmpty(String s) {
+    return s == null || s.isEmpty();
+  }
+
+
   /**
    * Check whether {@code name} is a valid kebab-case name.
    * A valid kebab-case name must satisfy the following conditions: